+++ FilterPanel.lua
@@ -114,7 +114,7 @@ local function addOptionToFilterFrame(parent, _, text, name)
 
 	filterOption.Name:SetText(text)
 
-	miog.FilterPanel.FilterOptions[name] = filterOption
+	parent[name] = filterOption
 
 	return filterOption
 end
@@ -122,18 +122,16 @@ end
 miog.addOptionToFilterFrame = addOptionToFilterFrame
 
 local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
-	local filterPanel = miog.FilterPanel
-
 	local minName = "min" .. name
 	local maxName = "max" .. name
 
 	local filterOption = CreateFrame("Frame", nil, parent, "MIOG_FilterOptionDualSpinnerTemplate")
-	filterOption:SetSize(parent:GetWidth(), 25)
+	filterOption:SetSize(220, 25)
 	filterOption.Button:HookScript("OnClick", function(self)
 		local currentPanel = LFGListFrame.activePanel:GetDebugName()
 		local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 	
-		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]["filterFor" .. name] = self:GetChecked()
+		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][name] = self:GetChecked()
 
 		convertFiltersToAdvancedBlizzardFilters()
 
@@ -148,14 +146,15 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 
 	filterOption.Name:SetText(text or name)
 
-	filterPanel.FilterOptions["filterFor" .. name] = filterOption
-	filterPanel.FilterOptions["linked" .. name] = filterOption.Link
+	parent[name] = filterOption
+	parent["Linked" .. name] = filterOption.Link
+
 
 	filterOption.Link:SetScript("OnClick", function(self)
 		local currentPanel = LFGListFrame.activePanel:GetDebugName()
 		local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 	
-		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]["linked" .. name] = self:GetChecked()
+		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]["Linked" .. name] = self:GetChecked()
 
 		if(LFGListFrame.activePanel == LFGListFrame.SearchPanel) then
 			miog.checkSearchResultListForEligibleMembers()
@@ -169,10 +168,12 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 	for i = 1, 2, 1 do
 		local currentName = i == 1 and minName or maxName
 
-		filterOption["Spinner" .. i]:SetWidth(21)
-		filterOption["Spinner" .. i]:SetMinMaxValues(range ~= nil and range or 0, 40)
+		local currentSpinner = filterOption[i == 1 and "Minimum" or "Maximum"]
+
+		currentSpinner:SetWidth(21)
+		currentSpinner:SetMinMaxValues(range ~= nil and range or 0, 40)
 		
-		filterOption["Spinner" .. i]:SetScript("OnTextChanged", function(self, userInput)
+		currentSpinner:SetScript("OnTextChanged", function(self, userInput)
 			if(userInput) then
 				local spinnerValue = self:GetNumber()
 
@@ -186,15 +187,15 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 					self:SetValue(spinnerValue)
 
 					if(i == 1) then
-						if(filterPanel.FilterOptions[maxName]:GetValue() < spinnerValue) then
-							filterPanel.FilterOptions[maxName]:SetValue(spinnerValue)
+						if(parent[maxName]:GetValue() < spinnerValue) then
+							parent[maxName]:SetValue(spinnerValue)
 							MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][maxName] = spinnerValue
 
 						end
 
 					else
-						if(filterPanel.FilterOptions[minName]:GetValue() > spinnerValue) then
-							filterPanel.FilterOptions[minName]:SetValue(spinnerValue)
+						if(parent[minName]:GetValue() > spinnerValue) then
+							parent[minName]:SetValue(spinnerValue)
 							MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][minName] = spinnerValue
 
 						end
@@ -213,23 +214,23 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 			end
 		end)
 
-		filterOption["Spinner" .. i].DecrementButton:SetScript("OnMouseDown", function(self)
-			filterOption["Spinner" .. i]:Decrement()
+		currentSpinner.DecrementButton:SetScript("OnMouseDown", function(self)
+			currentSpinner:Decrement()
 
-			local spinnerValue = filterOption["Spinner" .. i]:GetValue()
+			local spinnerValue = currentSpinner:GetValue()
 
 			local currentPanel = LFGListFrame.activePanel:GetDebugName()
 			local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 
 			MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][currentName] = spinnerValue
 
-			if(filterPanel.FilterOptions[minName]:GetValue() > spinnerValue) then
-				filterPanel.FilterOptions[minName]:SetValue(spinnerValue)
+			if(parent[minName]:GetValue() > spinnerValue) then
+				parent[minName]:SetValue(spinnerValue)
 				MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][minName] = spinnerValue
 
 			end
 
-			filterOption["Spinner" .. i]:ClearFocus()
+			currentSpinner:ClearFocus()
 
 			convertFiltersToAdvancedBlizzardFilters()
 		
@@ -242,23 +243,23 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 			end
 		end)
 
-		filterOption["Spinner" .. i].IncrementButton:SetScript("OnMouseDown", function()
-			filterOption["Spinner" .. i]:Increment()
+		currentSpinner.IncrementButton:SetScript("OnMouseDown", function()
+			currentSpinner:Increment()
 
-			local spinnerValue = filterOption["Spinner" .. i]:GetValue()
+			local spinnerValue = currentSpinner:GetValue()
 
 			local currentPanel = LFGListFrame.activePanel:GetDebugName()
 			local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 
 			MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][currentName] = spinnerValue
 
-			if(filterPanel.FilterOptions[maxName]:GetValue() < spinnerValue) then
-				filterPanel.FilterOptions[maxName]:SetValue(spinnerValue)
+			if(parent[maxName]:GetValue() < spinnerValue) then
+				parent[maxName]:SetValue(spinnerValue)
 				MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][maxName] = spinnerValue
 
 			end
 
-			filterOption["Spinner" .. i]:ClearFocus()
+			currentSpinner:ClearFocus()
 
 			convertFiltersToAdvancedBlizzardFilters()
 		
@@ -270,8 +271,6 @@ local function addDualNumericSpinnerToFilterFrame(parent, name, text, range)
 	
 			end
 		end)
-
-		filterPanel.FilterOptions[currentName] = filterOption["Spinner" .. i]
 	end
 
 	return filterOption
@@ -281,17 +280,15 @@ end
 miog.addDualNumericSpinnerToFilterFrame = addDualNumericSpinnerToFilterFrame
 
 local function addDualNumericFieldsToFilterFrame(parent, name)
-	local filterPanel = miog.FilterPanel
-
 	local filterOption = CreateFrame("Frame", nil, parent, "MIOG_FilterOptionDualFieldTemplate")
-	filterOption:SetSize(parent:GetWidth(), 25)
+	filterOption:SetSize(220, 25)
 
 	--local optionButton = miog.createBasicFrame("persistent", "UICheckButtonTemplate", parent, miog.C.INTERFACE_OPTION_BUTTON_SIZE, miog.C.INTERFACE_OPTION_BUTTON_SIZE)
 	filterOption.Button:HookScript("OnClick", function(self)
 		local currentPanel = LFGListFrame.activePanel:GetDebugName()
 		local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 
-		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]["filterFor" .. name] = self:GetChecked()
+		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][name] = self:GetChecked()
 
 		convertFiltersToAdvancedBlizzardFilters()
 		
@@ -304,7 +301,7 @@ local function addDualNumericFieldsToFilterFrame(parent, name)
 		end
 	end)
 
-	filterPanel.FilterOptions["filterFor" .. name] = filterOption
+	parent[name] = filterOption
 
 	filterOption.Name:SetText(name)
 
@@ -314,12 +311,12 @@ local function addDualNumericFieldsToFilterFrame(parent, name)
 	for i = 1, 2, 1 do
 		local currentName = i == 1 and minName or maxName
 
-		local numericField = filterOption["Field" .. i]
-		numericField:SetAutoFocus(false)
-		numericField.autoFocus = false
-		numericField:SetNumeric(true)
-		numericField:SetMaxLetters(4)
-		numericField:HookScript("OnTextChanged", function(self, ...)
+		local currentField = filterOption[i == 1 and "Minimum" or "Maximum"]
+		currentField:SetAutoFocus(false)
+		currentField.autoFocus = false
+		currentField:SetNumeric(true)
+		currentField:SetMaxLetters(4)
+		currentField:HookScript("OnTextChanged", function(self, ...)
 			PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
 			
 			local text = tonumber(self:GetText())
@@ -330,7 +327,7 @@ local function addDualNumericFieldsToFilterFrame(parent, name)
 
 			convertFiltersToAdvancedBlizzardFilters()
 
-			if(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]["filterFor" .. name]) then
+			if(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][name]) then
 				if(LFGListFrame.activePanel == LFGListFrame.SearchPanel) then
 					miog.checkSearchResultListForEligibleMembers()
 		
@@ -341,9 +338,6 @@ local function addDualNumericFieldsToFilterFrame(parent, name)
 			end
 
 		end)
-		
-
-		filterPanel.FilterOptions[currentName] = numericField
 	end
 
 	return filterOption
@@ -353,7 +347,6 @@ end
 miog.addDualNumericFieldsToFilterFrame = addDualNumericFieldsToFilterFrame
 
 local function updateDungeonCheckboxes()
-	local filterPanel = miog.FilterPanel
 	local sortedSeasonDungeons = {}
 
 	local seasonGroups = C_LFGList.GetAvailableActivityGroups(GROUP_FINDER_CATEGORY_ID_DUNGEONS, bit.bor(Enum.LFGListFilter.CurrentSeason, Enum.LFGListFilter.PvE));
@@ -370,7 +363,7 @@ local function updateDungeonCheckboxes()
 		end)
 
 		for k, activityEntry in ipairs(sortedSeasonDungeons) do
-			local currentButton = filterPanel.FilterOptions.DungeonPanel.Buttons[k]
+			local currentButton = miog.FilterPanel.IndexedOptions.Dungeons.Buttons[k]
 
 			currentButton:HookScript("OnClick", function(self)
 				MIOG_SavedSettings.filterOptions.table["LFGListFrame.SearchPanel"][2].dungeons[activityEntry.groupFinderActivityGroupID] = self:GetChecked()
@@ -397,7 +390,6 @@ end
 miog.updateDungeonCheckboxes = updateDungeonCheckboxes
 
 local function updateRaidCheckboxes()
-	local filterPanel = miog.FilterPanel
 	local sortedExpansionRaids = {}
 	
 	local seasonGroups = C_LFGList.GetAvailableActivityGroups(3, bit.bor(Enum.LFGListFilter.CurrentExpansion, Enum.LFGListFilter.PvE));
@@ -414,7 +406,7 @@ local function updateRaidCheckboxes()
 		end)
 
 		for k, activityEntry in ipairs(sortedExpansionRaids) do
-			local currentButton = filterPanel.FilterOptions.RaidPanel.Buttons[k]
+			local currentButton = miog.FilterPanel.IndexedOptions.Raids.Buttons[k]
 
 			currentButton:HookScript("OnClick", function(self)
 				MIOG_SavedSettings.filterOptions.table["LFGListFrame.SearchPanel"][3].raids[activityEntry.groupFinderActivityGroupID] = self:GetChecked()
@@ -440,16 +432,18 @@ end
 
 miog.updateRaidCheckboxes = updateRaidCheckboxes
 
-local function addRaidCheckboxes()
-	local filterPanel = miog.FilterPanel
-	
-	local raidPanel = miog.createBasicFrame("persistent", "BackdropTemplate", filterPanel.CategoryOptions, filterPanel.CategoryOptions:GetWidth(), miog.C.INTERFACE_OPTION_BUTTON_SIZE * 3)
+local function addRaidCheckboxes(parent)	
+	local raidPanel = miog.createBasicFrame("persistent", "BackdropTemplate", parent, parent:GetWidth(), miog.C.INTERFACE_OPTION_BUTTON_SIZE * 3)
 	raidPanel.Buttons = {}
 
-	filterPanel.FilterOptions.RaidPanel = raidPanel
+	parent.Raids = raidPanel
+
+	local raidOptionsButton = addOptionToFilterFrame(raidPanel, nil, "Raid options", "filterForRaids")
+	raidOptionsButton:SetPoint("TOPLEFT", raidPanel, "TOPLEFT", 0, 0)
+	raidPanel.Option = raidOptionsButton
 
 	local raidPanelFirstRow = miog.createBasicFrame("persistent", "BackdropTemplate", raidPanel, raidPanel:GetWidth(), raidPanel:GetHeight() / 2)
-	raidPanelFirstRow:SetPoint("TOPLEFT", raidPanel, "TOPLEFT")
+	raidPanelFirstRow:SetPoint("TOPLEFT", raidOptionsButton, "BOTTOMLEFT")
 
 	local raidPanelSecondRow = miog.createBasicFrame("persistent", "BackdropTemplate", raidPanel, raidPanel:GetWidth(), raidPanel:GetHeight() / 2)
 	raidPanelSecondRow:SetPoint("BOTTOMLEFT", raidPanel, "BOTTOMLEFT")
@@ -480,19 +474,21 @@ end
 
 miog.addRaidCheckboxes = addRaidCheckboxes
 
-local function addDungeonCheckboxes()
-	local filterPanel = miog.FilterPanel
-	
-	local dungeonPanel = miog.createBasicFrame("persistent", "BackdropTemplate", filterPanel.CategoryOptions, filterPanel.CategoryOptions:GetWidth(), miog.C.INTERFACE_OPTION_BUTTON_SIZE * 3)
+local function addDungeonCheckboxes(parent)	
+	local dungeonPanel = miog.createBasicFrame("persistent", "BackdropTemplate", parent, 220, 72)
 	dungeonPanel.Buttons = {}
 
-	filterPanel.FilterOptions.DungeonPanel = dungeonPanel
+	parent.Dungeons = dungeonPanel
 
-	local dungeonPanelFirstRow = miog.createBasicFrame("persistent", "BackdropTemplate", dungeonPanel, dungeonPanel:GetWidth(), dungeonPanel:GetHeight() / 2)
-	dungeonPanelFirstRow:SetPoint("TOPLEFT", dungeonPanel, "TOPLEFT")
+	local dungeonOptionsButton = addOptionToFilterFrame(dungeonPanel, nil, "Dungeon options", "filterForDungeons")
+	dungeonOptionsButton:SetPoint("TOPLEFT", dungeonPanel, "TOPLEFT", 0, 0)
+	dungeonPanel.Option = dungeonOptionsButton
 
-	local dungeonPanelSecondRow = miog.createBasicFrame("persistent", "BackdropTemplate", dungeonPanel, dungeonPanel:GetWidth(), dungeonPanel:GetHeight() / 2)
-	dungeonPanelSecondRow:SetPoint("BOTTOMLEFT", dungeonPanel, "BOTTOMLEFT")
+	local dungeonPanelFirstRow = miog.createBasicFrame("persistent", "BackdropTemplate", dungeonPanel, dungeonPanel:GetWidth(), 24)
+	dungeonPanelFirstRow:SetPoint("TOPLEFT", dungeonOptionsButton, "BOTTOMLEFT")
+
+	local dungeonPanelSecondRow = miog.createBasicFrame("persistent", "BackdropTemplate", dungeonPanel, dungeonPanel:GetWidth(), 24)
+	dungeonPanelSecondRow:SetPoint("TOPLEFT", dungeonPanelFirstRow, "BOTTOMLEFT")
 
 	local counter = 0
 
@@ -528,7 +524,7 @@ local function updateFilterDifficulties(reset)
 	local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 
 	if(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()]) then
-		local difficultyDropDown = miog.FilterPanel.FilterOptions.Dropdown
+		local difficultyDropDown = miog.FilterPanel.IndexedOptions.Difficulty.Dropdown
 		difficultyDropDown:ResetDropDown()
 
 		MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID] = MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID]
@@ -605,27 +601,35 @@ local function setupFiltersForActivePanel(reset)
 	local categoryID = currentPanel == "LFGListFrame.SearchPanel" and LFGListFrame.SearchPanel.categoryID or currentPanel == "LFGListFrame.ApplicationViewer" and C_LFGList.HasActiveEntryInfo() and C_LFGList.GetActivityInfoTable(C_LFGList.GetActiveEntryInfo().activityID).categoryID or LFGListFrame.CategorySelection.selectedCategory
 
 	if(currentPanel == "LFGListFrame.SearchPanel") then
-		miog.FilterPanel.SearchPanelOptions:Show()
-
-		if((categoryID == 2 or categoryID == 3)) then
-			miog.FilterPanel.CategoryOptions:Show()
-
-		else
-			miog.FilterPanel.CategoryOptions:Hide()
+		--miog.FilterPanel.SearchPanelOptions:Show()
 		
-		end
+		miog.FilterPanel.IndexedOptions.Score:SetShown(categoryID == 2 and true or false)
+		miog.FilterPanel.IndexedOptions.BossKills:SetShown(categoryID == 3 and true or false)
+		--miog.FilterPanel.IndexedOptions.Divider:SetShown((categoryID == 2 or categoryID == 3) and true or false)
+		miog.FilterPanel.IndexedOptions.Dungeons:SetShown(categoryID == 2 and true or false)
+		--miog.FilterPanel.IndexedOptions.filterForDungeons:SetShown(categoryID == 2 and true or false)
+		--miog.FilterPanel.IndexedOptions.RaidOptionsButton:SetShown(categoryID == 3 and true or false)
+		miog.FilterPanel.IndexedOptions.Raids:SetShown(categoryID == 3 and true or false)
+		--miog.FilterPanel.IndexedOptions.affixFitButton
+
 	else
-		miog.FilterPanel.SearchPanelOptions:Hide()
-		miog.FilterPanel.CategoryOptions:Hide()
+		
+		miog.FilterPanel.IndexedOptions.ScoreField:Hide()
+		miog.FilterPanel.IndexedOptions.BossKillsSpinner:Hide()
+		miog.FilterPanel.IndexedOptions.Divider:Hide()
+		miog.FilterPanel.IndexedOptions.DungeonOptionsButton:Hide()
+		miog.FilterPanel.IndexedOptions.DungeonPanel:Hide()
+		miog.FilterPanel.IndexedOptions.RaidOptionsButton:Hide()
+		miog.FilterPanel.IndexedOptions.RaidPanel:Hide()
 	
 	end
+
+	miog.FilterPanel.IndexedOptions:MarkDirty()
 
 
 	local isDungeon = categoryID == 2
 	local isRaid = categoryID == 3
 	local isPvp = categoryID == 4 or categoryID == 7
-
-	local filterPanel = miog.FilterPanel
 
 	if(categoryID) then
 		if(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID] == nil) then
@@ -636,46 +640,46 @@ local function setupFiltersForActivePanel(reset)
 		updateFilterDifficulties(reset)
 
 		for i = 1, 3, 1 do
-			filterPanel.FilterOptions.Roles.RoleButtons[i]:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForRoles[i == 1 and "TANK" or i == 2 and "HEALER" or "DAMAGER"])
+			miog.FilterPanel.IndexedOptions.Roles.Buttons[i]:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForRoles[i == 1 and "TANK" or i == 2 and "HEALER" or "DAMAGER"])
 		end
 
-		for k, v in pairs(filterPanel.FilterOptions) do
-			if(v.Button and v.Button:GetObjectType() == "CheckButton") then
-				v.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][k])
-			end
-		end
+		--for k, v in pairs(miog.FilterPanel.IndexedOptions) do
+		--	if(v.Button and v.Button:GetObjectType() == "CheckButton") then
+		--		v.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID][k])
+		--	end
+		--end
 
-		filterPanel.FilterOptions.filterForDifficulty:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDifficulty)
+		miog.FilterPanel.IndexedOptions.Difficulty:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDifficulty)
 
-		filterPanel.FilterOptions.linkedTanks:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedTanks)
-		filterPanel.FilterOptions.linkedHealers:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedHealers)
-		filterPanel.FilterOptions.linkedDamager:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedDamager)
+		miog.FilterPanel.IndexedOptions.LinkedTanks:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedTanks)
+		miog.FilterPanel.IndexedOptions.LinkedHealers:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedHealers)
+		miog.FilterPanel.IndexedOptions.LinkedTanks:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].linkedDamager)
 
-		filterPanel.FilterOptions.filterForTanks.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForTanks)
-		filterPanel.FilterOptions.minTanks:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minTanks)
-		filterPanel.FilterOptions.maxTanks:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxTanks)
+		miog.FilterPanel.IndexedOptions.Tanks.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForTanks)
+		miog.FilterPanel.IndexedOptions.Tanks.Minimum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minTanks)
+		miog.FilterPanel.IndexedOptions.Tanks.Maximum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxTanks)
 
-		filterPanel.FilterOptions.filterForHealers.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForHealers)
-		filterPanel.FilterOptions.minHealers:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minHealers)
-		filterPanel.FilterOptions.maxHealers:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxHealers)
+		miog.FilterPanel.IndexedOptions.Healers.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForHealers)
+		miog.FilterPanel.IndexedOptions.Healers.Minimum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minHealers)
+		miog.FilterPanel.IndexedOptions.Healers.Maximum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxHealers)
 
-		filterPanel.FilterOptions.filterForDamager.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDamager)
-		filterPanel.FilterOptions.minDamager:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minDamager)
-		filterPanel.FilterOptions.maxDamager:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxDamager)
+		miog.FilterPanel.IndexedOptions.Damager.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDamager)
+		miog.FilterPanel.IndexedOptions.Damager.Minimum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minDamager)
+		miog.FilterPanel.IndexedOptions.Damager.Maximum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxDamager)
 
-		filterPanel.FilterOptions.filterForScore.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForScore)
-		filterPanel.FilterOptions.minScore:SetNumber(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minScore)
-		filterPanel.FilterOptions.maxScore:SetNumber(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxScore)
+		miog.FilterPanel.IndexedOptions.Score.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForScore)
+		miog.FilterPanel.IndexedOptions.Score.Minimum:SetNumber(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minScore)
+		miog.FilterPanel.IndexedOptions.Score.Maximum:SetNumber(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxScore)
 
-		filterPanel.FilterOptions.filterForBossKills.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForBossKills)
-		filterPanel.FilterOptions.minBossKills:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minBossKills)
-		filterPanel.FilterOptions.maxBossKills:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxBossKills)
+		miog.FilterPanel.IndexedOptions.BossKills.Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForBossKills)
+		miog.FilterPanel.IndexedOptions.BossKills.Minimum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].minBossKills)
+		miog.FilterPanel.IndexedOptions.BossKills.Maximum:SetValue(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].maxBossKills)
 
 		--filterPanel.FilterOptions["filterForClassSpecs"].Button:SetChecked(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForClassSpecs)
-		filterPanel.FilterOptions.filterForClassSpecs.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForClassSpecs)
+		miog.FilterPanel.ClassSpecOptions.Option.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForClassSpecs)
 
 		for classIndex, classEntry in ipairs(miog.CLASSES) do
-			local currentClassPanel = filterPanel.FilterOptions.ClassPanels[classIndex]
+			local currentClassPanel = miog.FilterPanel.ClassSpecOptions.Panels[classIndex]
 
 			if(MIOG_SavedSettings and MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].classSpec) then
 				if(MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].classSpec.class[classIndex] ~= nil) then
@@ -793,7 +797,7 @@ local function setupFiltersForActivePanel(reset)
 			end
 		end
 		if(categoryID == 2) then
-			filterPanel.FilterOptions.filterForDungeons.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDungeons)
+			miog.FilterPanel.IndexedOptions.Dungeons.Option.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForDungeons)
 
 			local sortedSeasonDungeons = {}
 
@@ -812,12 +816,12 @@ local function setupFiltersForActivePanel(reset)
 
 				for k, activityEntry in ipairs(sortedSeasonDungeons) do
 					local notChecked = MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][2].dungeons[activityEntry.groupFinderActivityGroupID] == false
-					filterPanel.FilterOptions.DungeonPanel.Buttons[k]:SetChecked(reset or not notChecked)
+					miog.FilterPanel.IndexedOptions.Dungeons.Buttons[k]:SetChecked(reset or not notChecked)
 				end
 			end
 		elseif(categoryID == 3) then
 			local sortedExpansionRaids = {}
-			filterPanel.FilterOptions.filterForRaids.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForRaids)
+			miog.FilterPanel.IndexedOptions.Raids.Option.Button:SetChecked(reset or MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][categoryID].filterForRaids)
 
 			if(LFGListFrame.SearchPanel.filters == Enum.LFGListFilter.Recommended) then
 				for k, v in pairs(miog.ACTIVITY_INFO) do
@@ -833,62 +837,71 @@ local function setupFiltersForActivePanel(reset)
 
 				for k, activityEntry in ipairs(sortedExpansionRaids) do
 					local checked = MIOG_SavedSettings.filterOptions.table[LFGListFrame.activePanel:GetDebugName()][3].raids[activityEntry.groupFinderActivityGroupID]
-					filterPanel.FilterOptions.RaidPanel.Buttons[k]:SetChecked(reset or checked)
+					miog.FilterPanel.IndexedOptions.Raids.Buttons[k]:SetChecked(reset or checked)
 				end
 			end
 		end
 	end
 
+
+	miog.FilterPanel.IndexedOptions:MarkDirty()
 	miog.FilterPanel:MarkDirty()
 end
 
 miog.setupFiltersForActivePanel = setupFiltersForActivePanel
 
 local function createClassSpecFilters(parent)
-	miog.FilterPanel.FilterOptions.ClassPanels = {}
+	--miog.FilterPanel.FilterOptions.ClassPanels = {}
+	local classSpecOption = addOptionToFilterFrame(parent, nil, "Class / spec", "filterForClassSpecs")
+	classSpecOption.layoutIndex = 1
+	miog.FilterPanel.ClassSpecOptions.Option = classSpecOption
+	--classSpecOption:SetPoint("TOPLEFT", miog.FilterPanel.ClassSpecOptions, "TOPLEFT")
+
+	parent.Panels = {}
 
 	for classIndex, classEntry in ipairs(miog.CLASSES) do
 		local classPanel = CreateFrame("Frame", nil, parent, "MIOG_ClassSpecFilterRowTemplate")
-
-		classPanel:SetSize(parent:GetWidth(), 20)
-
-		if(classIndex > 1) then
-			classPanel:SetPoint("TOPLEFT", miog.FilterPanel.FilterOptions.ClassPanels[classIndex-1] or parent, miog.FilterPanel.FilterOptions.ClassPanels[classIndex-1] and "BOTTOMLEFT" or "TOPLEFT", 0, -1)
-		end
+		classPanel.layoutIndex = classIndex + 1
+		classPanel:SetWidth(parent:GetWidth())
 
 		local r, g, b = GetClassColor(classEntry.name)
 		miog.createFrameBorder(classPanel, 1, r, g, b, 0.9)
 		classPanel:SetBackdropColor(r, g, b, 0.6)
 
 		classPanel.Class.Texture:SetTexture(miog.C.STANDARD_FILE_PATH .. "/classIcons/" .. classEntry.name .. "_round.png")
+		classPanel.Class.rightPadding = 8
 
-		miog.FilterPanel.FilterOptions.ClassPanels[classIndex] = classPanel
+		--miog.FilterPanel.FilterOptions.ClassPanels[classIndex] = classPanel
+		parent.Panels[classIndex] = classPanel
 		classPanel.SpecFrames = {}
 
 		for specIndex, specID in pairs(classEntry.specs) do
 			local specEntry = miog.SPECIALIZATIONS[specID]
 
-			local currentSpec = CreateFrame("Frame", nil, parent, "MIOG_ClassSpecSingleOptionTemplate")
-			currentSpec:SetSize(36, 20)
-			currentSpec:SetPoint("LEFT", classPanel.SpecFrames[classEntry.specs[specIndex-1]] or classPanel.Class.Texture, "RIGHT", 8, 0)
+			local currentSpec = CreateFrame("Frame", nil, classPanel, "MIOG_ClassSpecSingleOptionTemplate")
+			currentSpec.layoutIndex = specIndex + 1
+			--currentSpec:SetSize(36, 20)
+			currentSpec.rightPadding = 8
 			currentSpec.Texture:SetTexture(specEntry.icon)
 
 			classPanel.SpecFrames[specID] = currentSpec
 
 		end
-
 	end
 
-	return miog.FilterPanel.FilterOptions.ClassPanels[1], miog.FilterPanel.FilterOptions.ClassPanels[#miog.CLASSES]
+	miog.FilterPanel.IndexedOptions:MarkDirty()
+	miog.FilterPanel:MarkDirty()
+
+	return parent.Panels[1], parent.Panels[#parent.Panels]
 end
 
 local function addRolePanel(parent)
 	local roleFilterPanel = miog.createBasicFrame("persistent", "BackdropTemplate", parent, 150, 20)
-	roleFilterPanel.RoleButtons = {}
+	roleFilterPanel.Buttons = {}
 
 	local lastTexture = nil
 
-	miog.FilterPanel.FilterOptions.Roles = roleFilterPanel
+	parent.Roles = roleFilterPanel
 
 	for i = 1, 3, 1 do
 		local toggleRoleButton = miog.createBasicFrame("persistent", "UICheckButtonTemplate", roleFilterPanel, miog.C.APPLICANT_MEMBER_HEIGHT, miog.C.APPLICANT_MEMBER_HEIGHT)
@@ -896,7 +909,7 @@ local function addRolePanel(parent)
 		toggleRoleButton:SetPushedAtlas("checkbox-minimal")
 		toggleRoleButton:SetCheckedTexture("checkmark-minimal")
 		toggleRoleButton:SetDisabledCheckedTexture("checkmark-minimal-disabled")
-		toggleRoleButton:SetPoint("LEFT", lastTexture or roleFilterPanel, lastTexture and "RIGHT" or "LEFT", lastTexture and 10 or 0, 0)
+		toggleRoleButton:SetPoint("LEFT", lastTexture or roleFilterPanel, lastTexture and "RIGHT" or "LEFT", lastTexture and 7 or 0, 0)
 		toggleRoleButton:RegisterForClicks("LeftButtonDown")
 
 		local roleTexture = miog.createBasicTexture("persistent", nil, roleFilterPanel, miog.C.APPLICANT_MEMBER_HEIGHT - 3, miog.C.APPLICANT_MEMBER_HEIGHT - 3, "ARTWORK")
@@ -943,11 +956,14 @@ local function addRolePanel(parent)
 
 		end
 
-		roleFilterPanel.RoleButtons[i] = toggleRoleButton
+		roleFilterPanel.Buttons[i] = toggleRoleButton
 
 		lastTexture = roleTexture
 
 	end
+
+	miog.FilterPanel.IndexedOptions:MarkDirty()
+	miog.FilterPanel:MarkDirty()
 
 	return roleFilterPanel
 end
@@ -955,7 +971,6 @@ end
 miog.loadFilterPanel = function()
 	miog.FilterPanel = CreateFrame("Frame", "MythicIOGrabber_FilterPanel", miog.Plugin, "MIOG_FilterPanel") ---@class Frame
 	miog.FilterPanel:SetPoint("TOPLEFT", miog.MainFrame, "TOPRIGHT", 5, 0)
-	miog.FilterPanel:SetSize(230, 650)
 	miog.FilterPanel:Hide()
 	miog.createFrameBorder(miog.FilterPanel, 1, CreateColorFromHexString(miog.C.BACKGROUND_COLOR_3):GetRGBA())
 	
@@ -1009,36 +1024,36 @@ miog.loadFilterPanel = function()
 		miog.FilterPanel:Hide()
 	end)
 
+	local firstClassPanel, lastClassPanel = createClassSpecFilters(miog.FilterPanel.ClassSpecOptions)
+	--firstClassPanel:SetPoint("TOPLEFT", classSpecOption, "BOTTOMLEFT")
 
-	miog.FilterPanel.FilterOptions = {}
-
-	local rolePanel = addRolePanel(miog.FilterPanel.StandardOptions)
-	rolePanel:SetPoint("TOPLEFT", miog.FilterPanel.StandardOptions, "TOPLEFT")
+	local rolePanel = addRolePanel(miog.FilterPanel.IndexedOptions)
+	rolePanel.layoutIndex = 1
 	
-	local classSpecOption = addOptionToFilterFrame(miog.FilterPanel.StandardOptions, nil, "Class / spec", "filterForClassSpecs")
-	classSpecOption:SetPoint("TOPLEFT", rolePanel, "BOTTOMLEFT")
-
-	local firstClassPanel, lastClassPanel = createClassSpecFilters(miog.FilterPanel.StandardOptions)
-	firstClassPanel:SetPoint("TOPLEFT", classSpecOption, "BOTTOMLEFT")
+	local partyFitButton = addOptionToFilterFrame(miog.FilterPanel.IndexedOptions, nil, "Party fit", "partyFit")
+	partyFitButton.layoutIndex = 2
+	--partyFitButton:SetPoint("TOPLEFT", lastClassPanel, "BOTTOMLEFT", 0, 0)
 	
-	local partyFitButton = addOptionToFilterFrame(miog.FilterPanel.StandardOptions, nil, "Party fit", "partyFit")
-	partyFitButton:SetPoint("TOPLEFT", lastClassPanel, "BOTTOMLEFT", 0, 0)
-	
-	local ressFitButton = addOptionToFilterFrame(miog.FilterPanel.StandardOptions, nil, "Ress fit", "ressFit")
-	ressFitButton:SetPoint("TOPLEFT", partyFitButton, "BOTTOMLEFT", 0, 0)
+	local ressFitButton = addOptionToFilterFrame(miog.FilterPanel.IndexedOptions, nil, "Ress fit", "ressFit")
+	ressFitButton.layoutIndex = 3
+	--ressFitButton:SetPoint("TOPLEFT", partyFitButton, "BOTTOMLEFT", 0, 0)
 
-	local lustFitButton = addOptionToFilterFrame(miog.FilterPanel.StandardOptions, nil, "Lust fit", "lustFit")
-	lustFitButton:SetPoint("TOPLEFT", ressFitButton, "BOTTOMLEFT", 0, 0)
+	local lustFitButton = addOptionToFilterFrame(miog.FilterPanel.IndexedOptions, nil, "Lust fit", "lustFit")
+	lustFitButton.layoutIndex = 4
+	--lustFitButton:SetPoint("TOPLEFT", ressFitButton, "BOTTOMLEFT", 0, 0)
+	--affixFitButton:SetPoint("TOPLEFT", lustFitButton, "BOTTOMLEFT", 0, 0)
 	
-	local hideHardDecline = addOptionToFilterFrame(miog.FilterPanel.SearchPanelOptions, nil, "Hide hard decline", "hardDecline")
-	hideHardDecline:SetPoint("TOPLEFT", miog.FilterPanel.SearchPanelOptions, "TOPLEFT", 0, 0)
+	local hideHardDecline = addOptionToFilterFrame(miog.FilterPanel.IndexedOptions, nil, "Hide hard decline", "hardDecline")
+	--hideHardDecline:SetPoint("TOPLEFT", miog.FilterPanel.SearchPanelOptions, "TOPLEFT", 0, 0)
+	hideHardDecline.layoutIndex = 5
 
-	local dropdownOptionButton = miog.createBasicFrame("persistent", "UICheckButtonTemplate", miog.FilterPanel.SearchPanelOptions, miog.C.INTERFACE_OPTION_BUTTON_SIZE, miog.C.INTERFACE_OPTION_BUTTON_SIZE)
+	local dropdownOptionButton = miog.createBasicFrame("persistent", "UICheckButtonTemplate", miog.FilterPanel.IndexedOptions, miog.C.INTERFACE_OPTION_BUTTON_SIZE, miog.C.INTERFACE_OPTION_BUTTON_SIZE)
 	dropdownOptionButton:SetNormalAtlas("checkbox-minimal")
 	dropdownOptionButton:SetPushedAtlas("checkbox-minimal")
 	dropdownOptionButton:SetCheckedTexture("checkmark-minimal")
 	dropdownOptionButton:SetDisabledCheckedTexture("checkmark-minimal-disabled")
-	dropdownOptionButton:SetPoint("TOPLEFT", hideHardDecline, "BOTTOMLEFT", 0, 0)
+	dropdownOptionButton.layoutIndex = 6
+	--dropdownOptionButton:SetPoint("TOPLEFT", hideHardDecline, "BOTTOMLEFT", 0, 0)
 	dropdownOptionButton:HookScript("OnClick", function(self)
 		PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
 
@@ -1055,47 +1070,54 @@ miog.loadFilterPanel = function()
 	
 			end
 	end)
-	miog.FilterPanel.FilterOptions.filterForDifficulty = dropdownOptionButton
+	miog.FilterPanel.IndexedOptions.Difficulty = dropdownOptionButton
 
-	local optionDropDown = Mixin(CreateFrame("Frame", nil, miog.FilterPanel.SearchPanelOptions, "MIOG_DropDownMenu"), SlickDropDown)
+	local optionDropDown = Mixin(CreateFrame("Frame", nil, miog.FilterPanel.IndexedOptions, "MIOG_DropDownMenu"), SlickDropDown)
 	optionDropDown:OnLoad()
-	optionDropDown:SetWidth(miog.FilterPanel:GetWidth() - dropdownOptionButton:GetWidth() - 5)
+	optionDropDown:SetWidth(miog.FilterPanel.fixedWidth - dropdownOptionButton:GetWidth() - 5)
 	optionDropDown:SetHeight(25)
-	optionDropDown:SetPoint("TOPLEFT", dropdownOptionButton, "TOPRIGHT")
+	optionDropDown:SetPoint("LEFT", dropdownOptionButton, "RIGHT")
 	optionDropDown:SetText("Choose a difficulty")
-	miog.FilterPanel.FilterOptions.Dropdown = optionDropDown
+	miog.FilterPanel.IndexedOptions.Difficulty.Dropdown = optionDropDown
 
-	local tanksSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.SearchPanelOptions, "Tanks")
-	tanksSpinner:SetPoint("TOPLEFT", dropdownOptionButton, "BOTTOMLEFT", 0, 0)
+	local tanksSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.IndexedOptions, "Tanks")
+	tanksSpinner.layoutIndex = 7
+	--tanksSpinner:SetPoint("TOPLEFT", dropdownOptionButton, "BOTTOMLEFT", 0, 0)
 
-	local healerSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.SearchPanelOptions, "Healers")
-	healerSpinner:SetPoint("TOPLEFT", tanksSpinner, "BOTTOMLEFT", 0, 0)
+	local healerSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.IndexedOptions, "Healers")
+	healerSpinner.layoutIndex = 8
+	--healerSpinner:SetPoint("TOPLEFT", tanksSpinner, "BOTTOMLEFT", 0, 0)
 
-	local damagerSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.SearchPanelOptions, "Damager")
-	damagerSpinner:SetPoint("TOPLEFT", healerSpinner, "BOTTOMLEFT", 0, 0)
+	local damagerSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.IndexedOptions, "Damager")
+	damagerSpinner.layoutIndex = 9
+	--damagerSpinner:SetPoint("TOPLEFT", healerSpinner, "BOTTOMLEFT", 0, 0)
 
-	local scoreField = miog.addDualNumericFieldsToFilterFrame(miog.FilterPanel.SearchPanelOptions, "Score")
-	scoreField:SetPoint("TOPLEFT", damagerSpinner, "BOTTOMLEFT", 0, 0)
+	local scoreField = miog.addDualNumericFieldsToFilterFrame(miog.FilterPanel.IndexedOptions, "Score")
+	scoreField.layoutIndex = 10
+	--scoreField:SetPoint("TOPLEFT", damagerSpinner, "BOTTOMLEFT", 0, 0)
 
-	local bossKillsSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.SearchPanelOptions, "BossKills", "Kills")
+	local bossKillsSpinner = miog.addDualNumericSpinnerToFilterFrame(miog.FilterPanel.IndexedOptions, "BossKills", "Kills")
+	bossKillsSpinner.layoutIndex = 11
 	bossKillsSpinner.Link:Hide()
-	bossKillsSpinner:SetPoint("TOPLEFT", damagerSpinner, "BOTTOMLEFT", 0, 0)
+	--bossKillsSpinner:SetPoint("TOPLEFT", damagerSpinner, "BOTTOMLEFT", 0, 0)
 
-	local divider = miog.createBasicTexture("persistent", nil, miog.FilterPanel.CategoryOptions, miog.FilterPanel.CategoryOptions:GetWidth(), 1, "BORDER")
+	local divider = miog.createBasicTexture("persistent", nil, miog.FilterPanel.IndexedOptions, miog.FilterPanel.fixedWidth, 1, "BORDER")
 	divider:SetAtlas("UI-LFG-DividerLine")
-	divider:SetPoint("BOTTOMLEFT", scoreField, "BOTTOMLEFT", 0, -5)
+	divider.layoutIndex = 12
 
-	local dungeonOptionsButton = addOptionToFilterFrame(miog.FilterPanel.CategoryOptions, nil, "Dungeon options", "filterForDungeons")
-	dungeonOptionsButton:SetPoint("TOPLEFT", divider, "BOTTOMLEFT", 0, 0)
+	local dungeonPanel = miog.addDungeonCheckboxes(miog.FilterPanel.IndexedOptions)
+	dungeonPanel.layoutIndex = 13
+	--dungeonPanel:SetPoint("TOPLEFT", dungeonOptionsButton, "BOTTOMLEFT", 0, 0)
 
-	local dungeonPanel = miog.addDungeonCheckboxes()
-	dungeonPanel:SetPoint("TOPLEFT", dungeonOptionsButton, "BOTTOMLEFT", 0, 0)
-	dungeonPanel.OptionsButton = dungeonOptionsButton
+	local raidPanel = miog.addRaidCheckboxes(miog.FilterPanel.IndexedOptions)
+	raidPanel.layoutIndex = 13
+	--raidPanel:SetPoint("TOPLEFT", raidOptionsButton, "BOTTOMLEFT", 0, 0)
+	--raidPanel.OptionsButton = raidOptionsButton
 
-	local raidOptionsButton = addOptionToFilterFrame(miog.FilterPanel.CategoryOptions, nil, "Raid options", "filterForRaids")
-	raidOptionsButton:SetPoint("TOPLEFT", divider, "BOTTOMLEFT", 0, 0)
+	--local affixFitButton = addOptionToFilterFrame(miog.FilterPanel.StandardOptions, nil, "Affix fit", "affixFit")
+	--affixFitButton.layoutIndex = 5
 
-	local raidPanel = miog.addRaidCheckboxes()
-	raidPanel:SetPoint("TOPLEFT", raidOptionsButton, "BOTTOMLEFT", 0, 0)
-	raidPanel.OptionsButton = raidOptionsButton
+	miog.FilterPanel.IndexedOptions:MarkDirty()
+	miog.FilterPanel:MarkDirty()
+	--divider:SetPoint("BOTTOMLEFT", scoreField, "BOTTOMLEFT", 0, -5)
 end